Remove all traces of vertica and use data API.
