+++ enterprise_reporting/clients/__init__.py
@@ -23,6 +23,8 @@ class EdxOAuth2APIClient(object):
     API_BASE_URL = LMS_ROOT_URL + '/api/'
     APPEND_SLASH = False
 
+    DEFAULT_VALUE_SAFEGUARD = object()
+
     def __init__(self, client_id=None, client_secret=None):
         """
         Connect to the REST API.
@@ -67,6 +69,47 @@ class EdxOAuth2APIClient(object):
                 self.connect()
             return func(self, *args, **kwargs)
         return inner
+
+    def _load_data(
+            self,
+            resource,
+            detail_resource=None,
+            resource_id=None,
+            querystring=None,
+            should_traverse_pagination=False,
+            default=DEFAULT_VALUE_SAFEGUARD,
+    ):
+        """
+        Loads a response from a call to one of the API endpoints.
+
+        Arguments:
+            resource: The endpoint resource name.
+            detail_resource: The sub-resource to append to the path.
+            resource_id: The resource ID for the specific detail to get from the endpoint.
+            querystring: Optional query string parameters.
+            should_traverse_pagination: Whether to traverse pagination or return paginated response.
+            default: The default value to return in case of no response content.
+
+        Returns
+            (JSON): Data returned by the API.
+        """
+        default_val = default if default is not self.DEFAULT_VALUE_SAFEGUARD else {}
+        querystring = querystring or {}
+
+        endpoint = getattr(self.client, resource)
+        endpoint = getattr(self.client, resource)(resource_id) if resource_id else endpoint
+        endpoint = getattr(endpoint, detail_resource) if detail_resource else endpoint
+        response = endpoint.get(**querystring)
+        if should_traverse_pagination:
+            results = traverse_pagination(response, endpoint)
+            response = {
+                'count': len(results),
+                'next': None,
+                'previous': None,
+                'results': results,
+            }
+
+        return response or default_val
 
 
 def traverse_pagination(response, endpoint):
