+++ enterprise_reporting/clients/enterprise.py
@@ -8,7 +8,7 @@ from __future__ import absolute_import, unicode_literals
 import os
 from collections import OrderedDict
 
-from enterprise_reporting.clients import EdxOAuth2APIClient, traverse_pagination
+from enterprise_reporting.clients import EdxOAuth2APIClient
 
 
 class EnterpriseAPIClient(EdxOAuth2APIClient):
@@ -24,8 +24,7 @@ class EnterpriseAPIClient(EdxOAuth2APIClient):
 
     PAGE_SIZE = os.environ.get('PAGE_SIZE', default=1000)
 
-    DEFAULT_VALUE_SAFEGUARD = object()
-
+    @EdxOAuth2APIClient.refresh_token
     def get_all_enterprise_reporting_configs(self, **kwargs):
         """
         Query the enterprise customer reporting endpoint for all available configs.
@@ -36,6 +35,7 @@ class EnterpriseAPIClient(EdxOAuth2APIClient):
             **kwargs
         )
 
+    @EdxOAuth2APIClient.refresh_token
     def get_enterprise_reporting_configs(self, enterprise_customer_uuid, **kwargs):
         """
         Query the enterprise customer reporting endpoint for a given enterprise customer.
@@ -47,6 +47,7 @@ class EnterpriseAPIClient(EdxOAuth2APIClient):
             **kwargs
         )
 
+    @EdxOAuth2APIClient.refresh_token
     def get_content_metadata(self, enterprise_customer_uuid):
         """Return all content metadata contained in the catalogs associated with an Enterprise Customer."""
         content_metadata = OrderedDict()
@@ -74,44 +75,23 @@ class EnterpriseAPIClient(EdxOAuth2APIClient):
         # We only made this a dictionary to help filter out duplicates by a common key. We just want values now.
         return content_metadata.values()
 
+
+class EnterpriseDataApiClient(EdxOAuth2APIClient):
+    """
+    Client for connecting to the Enterprise Data API.
+    """
+
+    API_BASE_URL = os.environ.get('ANALYTICS_API_URL') + '/enterprise/api/v0'
+    APPEND_SLASH = True
+
+    PAGE_SIZE = os.environ.get('PAGE_SIZE', default=1000)
+
     @EdxOAuth2APIClient.refresh_token
-    def _load_data(
-            self,
-            resource,
-            detail_resource=None,
-            resource_id=None,
-            querystring=None,
-            should_traverse_pagination=False,
-            default=DEFAULT_VALUE_SAFEGUARD,
-    ):
-        """
-        Loads a response from a call to one of the Enterprise endpoints.
-
-        Arguments:
-            resource: The endpoint resource name.
-            detail_resource: The sub-resource to append to the path.
-            resource_id: The resource ID for the specific detail to get from the endpoint.
-            querystring: Optional query string parameters.
-            should_traverse_pagination: Whether to traverse pagination or return paginated response.
-            default: The default value to return in case of no response content.
-
-        Returns
-            (JSON): Data returned by the API.
-        """
-        default_val = default if default is not self.DEFAULT_VALUE_SAFEGUARD else {}
-        querystring = querystring or {}
-
-        endpoint = getattr(self.client, resource)
-        endpoint = getattr(self.client, resource)(resource_id) if resource_id else endpoint
-        endpoint = getattr(endpoint, detail_resource) if detail_resource else endpoint
-        response = endpoint.get(**querystring)
-        if should_traverse_pagination:
-            results = traverse_pagination(response, endpoint)
-            response = {
-                'count': len(results),
-                'next': None,
-                'previous': None,
-                'results': results,
-            }
-
-        return response or default_val
+    def get_enterprise_enrollments(self, enterprise_customer_uuid):
+        return self._load_data(
+            'enterprise',
+            resource_id=enterprise_customer_uuid,
+            detail_resource='enrollments',
+            should_traverse_pagination=True,
+            querystring={'page_size': self.PAGE_SIZE},
+        )
