+++ requirements/base.txt
@@ -26,7 +26,6 @@ edx-drf-extensions==1.2.4
 edx-rest-api-client==1.7.1
 enum34==1.1.6             # via cryptography
 first==2.0.1              # via pip-tools
-future==0.16.0            # via vertica-python
 futures==3.2.0 ; python_version == "2.7"
 idna==2.6                 # via cryptography
 ipaddress==1.0.22         # via cryptography
@@ -41,8 +40,8 @@ pycparser==2.18           # via cffi
 pyjwt==1.6.1              # via djangorestframework-jwt, edx-rest-api-client
 pyminizip==0.2.1
 pynacl==1.2.1             # via paramiko
-python-dateutil==2.7.2    # via botocore, edx-drf-extensions, vertica-python
-pytz==2018.4              # via celery, django, vertica-python
+python-dateutil==2.7.3    # via botocore, edx-drf-extensions
+pytz==2018.4              # via celery, django
 pyyaml==3.12              # via awscli
 requests==2.9.1
 rsa==3.4.2                # via awscli
@@ -51,5 +50,4 @@ six==1.11.0
 slumber==0.7.1            # via edx-rest-api-client
 tox==3.0.0
 unicodecsv==0.14.1
-vertica-python==0.7.3
 virtualenv==15.2.0        # via tox
