+++ requirements/quality.in
@@ -1,9 +1,9 @@
 # Requirements for code quality checks
 
-caniusepython3            # Additional Python 3 compatibility pylint checks
-futures==3.2.0            # Used by caniusepython3, locked to allow install on python3
-edx-lint                  # edX pylint rules and plugins
-isort                     # to standardize order of imports
-pycodestyle               # PEP 8 compliance validation
-pydocstyle                # PEP 257 compliance validation
-testfixtures              # Mock objects for unit tests and doc tests
+caniusepython3                          # Additional Python 3 compatibility pylint checks
+futures==3.2.0; python_version == "2.7" # Used by caniusepython3, locked to allow install on python3
+edx-lint                                # edX pylint rules and plugins
+isort                                   # to standardize order of imports
+pycodestyle                             # PEP 8 compliance validation
+pydocstyle                              # PEP 257 compliance validation
+testfixtures                            # Mock objects for unit tests and doc tests
