+++ requirements/quality.txt
@@ -16,7 +16,7 @@ bcrypt==3.1.4             # via paramiko
 billiard==3.3.0.23        # via celery
 boto3==1.4.7
 botocore==1.7.36          # via awscli, boto3, s3transfer
-caniusepython3==6.0.0
+caniusepython3==7.0.0
 celery==3.1.18
 cffi==1.11.5              # via bcrypt, cryptography, pynacl
 click-log==0.1.8          # via edx-lint
@@ -26,11 +26,11 @@ configparser==3.5.0       # via pylint
 cookies==2.2.1            # via responses
 coverage==4.5.1           # via pytest-cov
 cryptography==1.9
-ddt==1.1.2
+ddt==1.1.3
 diff-cover==1.0.2
 distlib==0.2.7            # via caniusepython3
 django-fernet-fields==0.5
-django-model-utils==3.1.1
+django-model-utils==3.1.2
 django==1.11.5
 djangorestframework-jwt==1.11.0  # via edx-drf-extensions
 djangorestframework==3.6.3  # via edx-drf-extensions
@@ -41,15 +41,14 @@ edx-lint==0.5.5
 edx-rest-api-client==1.7.1
 enum34==1.1.6             # via astroid, cryptography
 factory-boy==2.11.1
-faker==0.8.13             # via factory-boy
+faker==0.8.15             # via factory-boy
 first==2.0.1              # via pip-tools
 flaky==3.4.0
 freezegun==0.3.10
 funcsigs==1.0.2           # via mock, pytest
-future==0.16.0            # via vertica-python
 futures==3.2.0 ; python_version == "2.7"
 idna==2.6                 # via cryptography
-inflect==0.2.5            # via jinja2-pluralize
+inflect==0.3.1            # via jinja2-pluralize
 ipaddress==1.0.22         # via cryptography, faker
 isort==4.3.4
 jinja2-pluralize==0.3.0   # via diff-cover
@@ -64,7 +63,7 @@ more-itertools==4.1.0     # via pytest
 packaging==17.1           # via caniusepython3
 paramiko==2.4
 path.py==11.0.1           # via edx-i18n-tools
-pbr==4.0.2                # via mock
+pbr==4.0.3                # via mock
 pip-tools==2.0.2
 pkginfo==1.4.2            # via twine
 pluggy==0.6.0             # via pytest, tox
@@ -87,8 +86,8 @@ pytest-catchlog==1.2.2
 pytest-cov==2.5.1
 pytest-django==3.2.1
 pytest==3.5.1             # via pytest-catchlog, pytest-cov, pytest-django
-python-dateutil==2.7.2    # via botocore, edx-drf-extensions, faker, freezegun, vertica-python
-pytz==2018.4              # via celery, django, vertica-python
+python-dateutil==2.7.3    # via botocore, edx-drf-extensions, faker, freezegun
+pytz==2018.4              # via celery, django
 pyyaml==3.12              # via awscli, edx-i18n-tools
 requests-toolbelt==0.8.0  # via twine
 requests==2.9.1
@@ -103,10 +102,9 @@ testfixtures==6.0.2
 text-unidecode==1.2       # via faker
 tox-battery==0.5
 tox==3.0.0
-tqdm==4.23.2              # via twine
+tqdm==4.23.3              # via twine
 twine==1.11.0
 unicodecsv==0.14.1
-vertica-python==0.7.3
 virtualenv==15.2.0        # via tox
-wheel==0.31.0
+wheel==0.31.1
 wrapt==1.10.11            # via astroid
