+++ requirements/test-master.in
@@ -1,3 +1,3 @@
 django==1.11.5                          # Application server
-djangorestframework==3.6.2              # REST API extensions for Django
-edx-drf-extensions==1.2.3               # edX extensions to django rest framework
+djangorestframework==3.6.3              # REST API extensions for Django
+edx-drf-extensions==1.2.4               # edX extensions to django rest framework
