+++ requirements/test-master.txt
@@ -7,68 +7,67 @@
 amqp==1.4.9               # via kombu
 anyjson==0.3.3            # via kombu
 asn1crypto==0.24.0        # via cryptography
-attrs==17.4.0             # via pytest
+attrs==18.1.0             # via pytest
 awscli==1.11.178
 bcrypt==3.1.4             # via paramiko
 billiard==3.3.0.23        # via celery
 boto3==1.4.7
 botocore==1.7.36          # via awscli, boto3, s3transfer
 celery==3.1.18
-cffi==1.11.4              # via bcrypt, cryptography, pynacl
+cffi==1.11.5              # via bcrypt, cryptography, pynacl
 click==6.7                # via pip-tools
 colorama==0.3.7           # via awscli
 cookies==2.2.1            # via responses
 coverage==4.5.1           # via pytest-cov
 cryptography==1.9
-ddt==1.1.1
+ddt==1.1.3
 django-fernet-fields==0.5
-django-model-utils==3.1.1
+django-model-utils==3.1.2
 django==1.11.5
 djangorestframework-jwt==1.11.0  # via edx-drf-extensions
-djangorestframework==3.6.2
+djangorestframework==3.6.3
 docutils==0.14            # via awscli, botocore
-edx-drf-extensions==1.2.3
+edx-drf-extensions==1.2.4
 edx-rest-api-client==1.7.1
 enum34==1.1.6             # via cryptography
-factory-boy==2.10.0
-faker==0.8.11             # via factory-boy
+factory-boy==2.11.1
+faker==0.8.15             # via factory-boy
 first==2.0.1              # via pip-tools
 flaky==3.4.0
-freezegun==0.3.9
+freezegun==0.3.10
 funcsigs==1.0.2           # via mock, pytest
-future==0.16.0            # via vertica-python
 futures==3.2.0 ; python_version == "2.7"
 idna==2.6                 # via cryptography
-ipaddress==1.0.19         # via cryptography, faker
+ipaddress==1.0.22         # via cryptography, faker
 jmespath==0.9.3           # via boto3, botocore
 kombu==3.0.37             # via celery
 mock==2.0.0
+more-itertools==4.1.0     # via pytest
 paramiko==2.4
-pbr==3.1.1                # via mock
-pip-tools==1.11.0
+pbr==4.0.3                # via mock
+pip-tools==2.0.2
 pluggy==0.6.0             # via pytest, tox
-py==1.5.2                 # via pytest, pytest-catchlog, tox
+py==1.5.3                 # via pytest, pytest-catchlog, tox
 pyasn1==0.4.2             # via paramiko, rsa
 pycparser==2.18           # via cffi
-pyjwt==1.5.3              # via djangorestframework-jwt, edx-rest-api-client
+pyjwt==1.6.1              # via djangorestframework-jwt, edx-rest-api-client
 pyminizip==0.2.1
 pynacl==1.2.1             # via paramiko
 pytest-catchlog==1.2.2
 pytest-cov==2.5.1
-pytest-django==3.1.2
-pytest==3.4.0             # via pytest-catchlog, pytest-cov, pytest-django
-python-dateutil==2.6.1    # via botocore, edx-drf-extensions, faker, freezegun, vertica-python
-pytz==2018.3              # via celery, django, vertica-python
+pytest-django==3.2.1
+pytest==3.5.1             # via pytest-catchlog, pytest-cov, pytest-django
+python-dateutil==2.7.3    # via botocore, edx-drf-extensions, faker, freezegun
+pytz==2018.4              # via celery, django
 pyyaml==3.12              # via awscli
 requests==2.9.1
-responses==0.8.1
+responses==0.9.0
 rsa==3.4.2                # via awscli
-s3transfer==0.1.12        # via awscli, boto3
-six==1.11.0               # via bcrypt, cryptography, faker, freezegun, mock, pip-tools, pynacl, pytest, python-dateutil, responses, tox, vertica-python
+s3transfer==0.1.13        # via awscli, boto3
+six==1.11.0
 slumber==0.7.1            # via edx-rest-api-client
-testfixtures==5.4.0
-text-unidecode==1.1       # via faker
-tox==2.9.1
+testfixtures==6.0.2
+text-unidecode==1.2       # via faker
+tox==3.0.0
 unicodecsv==0.14.1
-vertica-python==0.7.3
-virtualenv==15.1.0        # via tox
+virtualenv==15.2.0        # via tox
