+++ requirements/test-reporting.txt
@@ -10,7 +10,7 @@ ddt==1.1.2
 funcsigs==1.0.2           # via mock, pytest
 mock==2.0.0
 more-itertools==4.1.0     # via pytest
-pbr==4.0.2                # via mock
+pbr==4.0.3                # via mock
 pluggy==0.6.0             # via pytest, tox
 py==1.5.3                 # via pytest, tox
 pytest-cov==2.5.1
