+++ requirements/test.txt
@@ -20,9 +20,9 @@ colorama==0.3.7           # via awscli
 cookies==2.2.1            # via responses
 coverage==4.5.1           # via pytest-cov
 cryptography==1.9
-ddt==1.1.2
+ddt==1.1.3
 django-fernet-fields==0.5
-django-model-utils==3.1.1
+django-model-utils==3.1.2
 django==1.11.5
 djangorestframework-jwt==1.11.0  # via edx-drf-extensions
 djangorestframework==3.6.3  # via edx-drf-extensions
@@ -31,12 +31,11 @@ edx-drf-extensions==1.2.4
 edx-rest-api-client==1.7.1
 enum34==1.1.6             # via cryptography
 factory-boy==2.11.1
-faker==0.8.13             # via factory-boy
+faker==0.8.15             # via factory-boy
 first==2.0.1              # via pip-tools
 flaky==3.4.0
 freezegun==0.3.10
 funcsigs==1.0.2           # via mock, pytest
-future==0.16.0            # via vertica-python
 futures==3.2.0 ; python_version == "2.7"
 idna==2.6                 # via cryptography
 ipaddress==1.0.22         # via cryptography, faker
@@ -45,7 +44,7 @@ kombu==3.0.37             # via celery
 mock==2.0.0
 more-itertools==4.1.0     # via pytest
 paramiko==2.4
-pbr==4.0.2                # via mock
+pbr==4.0.3                # via mock
 pip-tools==2.0.2
 pluggy==0.6.0             # via pytest, tox
 py==1.5.3                 # via pytest, pytest-catchlog, tox
@@ -58,8 +57,8 @@ pytest-catchlog==1.2.2
 pytest-cov==2.5.1
 pytest-django==3.2.1
 pytest==3.5.1             # via pytest-catchlog, pytest-cov, pytest-django
-python-dateutil==2.7.2    # via botocore, edx-drf-extensions, faker, freezegun, vertica-python
-pytz==2018.4              # via celery, django, vertica-python
+python-dateutil==2.7.3    # via botocore, edx-drf-extensions, faker, freezegun
+pytz==2018.4              # via celery, django
 pyyaml==3.12              # via awscli
 requests==2.9.1
 responses==0.9.0
@@ -71,5 +70,4 @@ testfixtures==6.0.2
 text-unidecode==1.2       # via faker
 tox==3.0.0
 unicodecsv==0.14.1
-vertica-python==0.7.3
 virtualenv==15.2.0        # via tox
